Dialog code refactor
