+++ .vscode/settings.json
@@ -2,6 +2,7 @@
   "typescript.tsdk": "node_modules/typescript/lib",
   "cSpell.words": [
     "ccrte",
-    "inlines"
+    "inlines",
+    "strikethrough"
   ]
 }
