+++ src/index.tsx
@@ -1,6 +1,14 @@
 export { DocumentJSON } from "slate";
 export { Editor } from "slate-react";
 export { EditorContent, EditorRange, EditorValue, EFormat, slateToText, textToSlate } from "./common/slate-types";
+export { DialogContent } from "./modal-dialog/dialog-content";
+export { DialogFooter } from "./modal-dialog/dialog-footer";
+export { DialogHeader } from "./modal-dialog/dialog-header";
+export {
+  DisplayDialogSettings, FieldType, IDialogController, IField, IFieldValues, IRow, SelectOptions, SelectValue
+} from "./modal-dialog/dialog-types";
+export { ModalCover } from "./modal-dialog/modal-cover";
+export { ModalDialog } from "./modal-dialog/modal-dialog";
 export { HtmlSerializablePlugin } from "./plugins/html-serializable-plugin";
 export { htmlToSlate, slateToHtml } from "./serialization/html-serializer";
 export {
@@ -12,11 +20,19 @@ export {
 export { SlateContainer } from "./slate-container/slate-container";
 export { SlateEditor } from "./slate-editor/slate-editor";
 export { getContentHeight, handleToggleSuperSubscript } from "./slate-editor/slate-utils";
-export { DisplayDialogSettings, SlateToolbar, ToolbarTransform } from "./slate-toolbar/slate-toolbar";
+export { SlateToolbar, ToolbarTransform } from "./slate-toolbar/slate-toolbar";
 export { EditorToolbar, IButtonSpec, IToolbarColors, getPlatformTooltip } from "./editor-toolbar/editor-toolbar";
 export {
   OnChangeColorFn, OnChangeFn, OnClickFn, OnDidInvokeToolFn, OnMouseFn, ToolbarButton
 } from "./editor-toolbar/toolbar-button";
+
+import "./modal-dialog/dialog-content.scss";
+import "./modal-dialog/dialog-footer.scss";
+import "./modal-dialog/dialog-header.scss";
+import "./modal-dialog/form-checkbox-field.scss";
+import "./modal-dialog/form-field-entry.scss";
+import "./modal-dialog/modal-cover.scss";
+import "./modal-dialog/modal-dialog.scss";
 import "./slate-container/slate-container.scss";
 import "./slate-editor/slate-editor.scss";
 import "./editor-toolbar/editor-toolbar.scss";
