+++ src/modal-dialog/dialog-header.scss
@@ -1,0 +1,9 @@
+.ccrte-dialog-header {
+  height: 30px;
+  background-color: #A0A0A0;
+  color: white;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border-radius: 5px 5px 0 0;
+}
