+++ src/modal-dialog/dialog-vars.scss
@@ -1,0 +1,3 @@
+$dialogWidth: 512px;
+// height varies with content; this value is just used for placement
+$dialogHeight: 340px;
