+++ src/modal-dialog/form-checkbox-field.scss
@@ -1,0 +1,38 @@
+// cf. https://dev.to/proticm/styling-html-checkboxes-is-super-easy-302o
+.ccrte-checkbox-field {
+  display: inline-flex;
+  cursor: pointer;
+  position: relative;
+}
+
+.ccrte-checkbox-field > span {
+  padding: 0.5rem 0.25rem;
+  margin-top: -2px;
+}
+
+.ccrte-checkbox-field > input {
+  height: 20px !important;
+  width: 20px;
+  -webkit-appearance: none;
+  -moz-appearance: none;
+  -o-appearance: none;
+  appearance: none;
+  border: 1px solid #34495E;
+  border-radius: 4px;
+  outline: none;
+  transition-duration: 0.3s;
+  cursor: pointer;
+}
+
+.ccrte-checkbox-field > input:checked + span::before {
+  content: '\2713';
+  display: block;
+  text-align: center;
+  position: absolute;
+  left: 0.6rem;
+  top: 0.4rem;
+}
+
+.ccrte-checkbox-field > input:active {
+  border: 2px solid #34495E;
+}
