+++ src/modal-dialog/modal-cover.scss
@@ -1,0 +1,9 @@
+.ccrte-modal-cover {
+  width: 100%;
+  height: 100%;
+  background-color: black;
+  opacity: .01;
+  position: fixed;
+  top: 0;
+  left: 0;
+}
