+++ src/modal-dialog/modal-dialog.scss
@@ -1,0 +1,16 @@
+@import "./dialog-vars.scss";
+
+.ccrte-modal-dialog {
+  .ccrte-dialog {
+    position: fixed;
+    left: calc(50% - #{$dialogWidth} / 2);
+    top: calc(50% - #{$dialogHeight} / 2);
+    display: flex;
+    flex-direction: column;
+    width: $dialogWidth;
+    border: 2px solid #B3B3B3;
+    border-radius: 6px;
+    box-shadow: -2px 2px 5px grey;
+    background-color: white;
+  }
+}
