+++ src/plugins/image-plugin.tsx
@@ -5,9 +5,8 @@ import { isWebUri } from "valid-url";
 import { Inline } from "slate";
 import { Editor, RenderAttributes, RenderInlineProps } from "slate-react";
 import { EFormat, kSlateVoidClass } from "../common/slate-types";
+import { IDialogController, IFieldValues } from "../modal-dialog/dialog-types";
 import { hasActiveInline } from "../slate-editor/slate-utils";
-import { IFieldValues } from "../slate-toolbar/modal-dialog";
-import { IDialogController } from "../slate-toolbar/slate-toolbar";
 import { getDataFromElement, getRenderAttributesFromNode, classArray } from "../serialization/html-utils";
 import { HtmlSerializablePlugin } from "./html-serializable-plugin";
 import "./image-plugin.scss";
