--- src/slate-toolbar/modal-dialog.scss
@@ -1,131 +1,0 @@
-$dialogWidth: 512px;
-// height varies with content; this value is just used for placement
-$dialogHeight: 340px;
-
-.ccrte-modal-dialog {
-  .ccrte-modal-cover {
-    width: 100%;
-    height: 100%;
-    background-color: black;
-    opacity: .01;
-    position: fixed;
-    top: 0;
-    left: 0;
-  }
-  .dialog {
-    position: fixed;
-    left: calc(50% - #{$dialogWidth} / 2);
-    top: calc(50% - #{$dialogHeight} / 2);
-    display: flex;
-    flex-direction: column;
-    width: $dialogWidth;
-    border: 2px solid #B3B3B3;
-    border-radius: 6px;
-    box-shadow: -2px 2px 5px grey;
-    background-color: white;
-
-    .header {
-      height: 30px;
-      background-color: #A0A0A0;
-      color: white;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      border-radius: 5px 5px 0 0;
-    }
-    .content {
-      margin: 5px;
-
-      .inline-row {
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-      }
-
-      > .input-entry {
-        margin-top: 8px;
-      }
-
-      .input-entry {
-        &.inline {
-          display: inline-block;
-          &.checkbox {
-            margin-top: 3.5px;
-          }
-        }
-        .label {
-          margin: 5px;
-          color: black;
-        }
-        input {
-          margin: 5px;
-          height: 24px;
-          padding: 0 3px;
-
-          &.full-row {
-            width: $dialogWidth - 40px;
-          }
-        }
-
-        // cf. https://dev.to/proticm/styling-html-checkboxes-is-super-easy-302o
-        .checkbox {
-          display: inline-flex;
-          cursor: pointer;
-          position: relative;
-        }
-
-        .checkbox > span {
-          padding: 0.5rem 0.25rem;
-          margin-top: -2px;
-        }
-
-        .checkbox > input {
-          height: 20px;
-          width: 20px;
-          -webkit-appearance: none;
-          -moz-appearance: none;
-          -o-appearance: none;
-          appearance: none;
-          border: 1px solid #34495E;
-          border-radius: 4px;
-          outline: none;
-          transition-duration: 0.3s;
-          cursor: pointer;
-        }
-
-        .checkbox > input:checked + span::before {
-          content: '\2713';
-          display: block;
-          text-align: center;
-          position: absolute;
-          left: 0.6rem;
-          top: 0.4rem;
-        }
-
-        .checkbox > input:active {
-          border: 2px solid #34495E;
-        }
-      }
-    }
-    .footer {
-      margin: 5px;
-      display: flex;
-      justify-content: flex-end;
-    }
-    button {
-      height: 30px;
-      min-width: 80px;
-      border: 2px solid #dfdfdf;
-      background-color: #A0A0A0;
-      color: white;
-      cursor: pointer;
-      border-radius: 6px;
-      &:hover {
-        opacity: .9;
-      }
-      &:active {
-        opacity: .8;
-      }
-    }
-  }
-}
